// src/types/characteristic.types.ts
export interface CharacteristicCard {
  id: number;
  title: string;           // "Smart"
  price: string;           // "2 868"
  specs: string[];         // массив характеристик
  colors: string[];        // массив цветов
  image: string;           // путь к картинке
}